WSHUB-100: client create/update: Support `createRec` without `id`.
Second generation WideSky uses UUIDs for entity IDs, which can be
generated by the server itself, so it is no longer a requirement to
specify the entity's `id`.

We'll put `id`, `name` and `dis` in the first three columns, for
consistency, and sort the remainder by name.  Only `updateRec` will
require `id` to be given.