WC-5300: Normalise file headers
We have a few inconsistent headers with all the files in this project,
largely because of this project's heritage.

Header needs:
- text editor file format hints (e.g. vi modeline; these need to be the
  first few lines of the file in most cases)
- langauge-specific annotations (e.g. `"use strict"`, should proceed any
  significant code)
- Copyright notice (should proceed any significant code)
- License identifier: https://spdx.dev/ids/ (should accompany the
  copyright notice)

So let's add one, keep it short and sweet.  We provide the license
in the `LICENSE` file anyway, and the SPDX-standard format should mean
code analysis tools will spot that and recognise it as MIT anyway, there
is no benefit in duplicating the license text here.